requirements.txt
transformers
torch
requests
numpy
Pillow
